requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
tzdata==2024.1
